Move initLocalStorage to main.tsx and define variables for localStorage keys
